fixpipe scale value输入从uint64转fp32



# message auto-generated for no-merge-commit merge:
!1471 merge fixuint64 into master

fixpipe scale value输入从uint64转fp32

Created-by: wang-zhiyue2
Commit-by: wang-zhiyue2
Merged-by: turing_project1
Description: 描述：
fixpipe scale value从uint64转fp32输入
测试：
ST用例2/2精度通过
文档更新：
Ascend c++特性开发说明书
类型标签：
参数更改

See merge request: cann/pypto-dev!1471

+++ framework/include/tilefwk/tilefwk_op.h
@@ -365,10 +365,10 @@ enum class ReLuType : int64_t
 struct MatmulExtendParam {
     Tensor biasTensor{Tensor()};
     Tensor scaleTensor{Tensor()};
-    uint64_t scaleValue{0};
+    float scaleValue{0.0f};
     ReLuType reluType{ReLuType::NoReLu};
 
-    MatmulExtendParam(Tensor bias, Tensor scale, uint64_t scaleVal, ReLuType relu)
+    MatmulExtendParam(Tensor bias, Tensor scale, float scaleVal, ReLuType relu)
         : biasTensor(std::move(bias)),
           scaleTensor(std::move(scale)),
           scaleValue(scaleVal),
